// src/types/employee/EmployeePosition.ts
export enum EmployeePosition {
    OPERATOR = "OPERATOR",
    TECHNICIAN = "TECHNICIAN",
    SUPERVISOR = "SUPERVISOR",
    QUALITY_INSPECTOR = "QUALITY_INSPECTOR",
    PRODUCTION_MANAGER = "PRODUCTION_MANAGER",
    ASSEMBLER = "ASSEMBLER",
    MAINTENANCE_ENGINEER = "MAINTENANCE_ENGINEER",
    LOGISTICS_COORDINATOR = "LOGISTICS_COORDINATOR",
    SAFETY_OFFICER = "SAFETY_OFFICER",
    SHIFT_LEADER = "SHIFT_LEADER",
    TOOLMAKER = "TOOLMAKER",
}